optimisation des float vs double

+++ m_VCF.h
@@ -40,20 +40,20 @@ inline float VCF1(uint32_t j, float fq, float input1) { //, float res, float mod
     else
         mod_q  += g_pot_audio[k_VCF1_mod2] * g_Modulation[curent_config.c_Modulation_Source[VCF1_MOD2]];
 
-    float freq = fq+ (48.f*mod_fq) + (24.f*g_Modulation[MIDI_expression]);
+    double freq = fq+ (48.f*mod_fq) + (24.f*g_Modulation[MIDI_expression]);
     tmp = curent_config.c_VCF1_pitch_TRACK;
-    freq += tmp * 0.5 * (allvoice[j].v_pitch-12.f);
-    freq = _fclamp(freq, -128, 138);
+    freq += tmp * 0.5f * (allvoice[j].v_pitch-12.f);
+    freq = _fclamp(freq, -128.f, 138.f);
     freq = CV2freq(freq);
 
     float omega = freq * (TWOPI_F/48000.f);
-    float g = omega * (0.9892f + omega*(-0.4342f + omega*(0.1381f + omega * -0.0202f)));
+    float g = omega * (0.9892 + omega*(-0.4342 + omega*(0.1381 + omega * -0.0202)));
 
-    float Q = 4.5f * _fclamp(res + mod_q, 0.f, 1.f);
+    float Q = 4.5f * _fclamp(res + mod_q, 0., 1.);
     // resonance frequency compensation
     Q *= 1.0029f + omega*(0.0526f + omega * (-0.0926f  + omega*0.0218f));
 
-    input1 *= 0.5; // limitation de l'amplitude d'entrée pour ne pas trop distordre le signal avant le filtre
+    input1 *= 0.5f; // limitation de l'amplitude d'entrée pour ne pas trop distordre le signal avant le filtre
 
 	Q *= 1.01f; // ??? c'est plus lent si je vire cette ligne!!!
